% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosense.R
\name{rsim.sense.path}
\alias{rsim.sense.path}
\title{Ecosense function for rpath (rsim.sense.path)}
\usage{
rsim.sense.path(
  Rpath.scenario,
  Rpath,
  Rpath.params,
  steps_yr = 12,
  steps_m = 1
)
}
\value{
Returns an Rsim.scenario object that can be supplied to the rsim.run function.
}
\description{
5 November 2019
}
\seealso{
Other Rpath functions

Whitehouse and Aydin (Submitted) Assessing the sensitivity of three Alaska marine
food webs to perturbations: an example of Ecosim simulations using Rpath

This function generates random parameters around the Ecopath baseline,
not the scenario.: 
\code{\link{rsim.sense}()}
}
\concept{Rpath functions

Whitehouse and Aydin (Submitted) Assessing the sensitivity of three Alaska marine
food webs to perturbations: an example of Ecosim simulations using Rpath

This function generates random parameters around the Ecopath baseline,
not the scenario.}
