% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpath_support.R
\name{RsimEndpoints}
\alias{RsimEndpoints}
\alias{starting.biomass}
\alias{ending.biomass}
\alias{starting.catch}
\alias{ending.catch}
\title{Extract start and end states from an rsim run output.}
\usage{
starting.biomass(Rsim.output)

ending.biomass(Rsim.output)

starting.catch(Rsim.output)

ending.catch(Rsim.output)
}
\arguments{
\item{Rsim.output}{An Rsim output object generated by rsim.run or rsim.step.}
}
\value{
Returns a named vector of starting or ending biomass or catch.
}
\description{
Extract the starting or ending biomass or catch from the output of an rsim run.
}
\seealso{
Other Rpath functions: 
\code{\link{MTI}()},
\code{\link{adjust.fishing}()},
\code{\link{adjust.forcing}()},
\code{\link{adjust.scenario}()},
\code{\link{check.rpath.params}()},
\code{\link{create.rpath.params}()},
\code{\link{extract.node}()},
\code{\link{frate.table}()},
\code{\link{read.rpath.params}()},
\code{\link{rpath.groups}()},
\code{\link{rpath.stanzas}()},
\code{\link{rpath}()},
\code{\link{rsim.params}()},
\code{\link{rsim.plot}()},
\code{\link{rsim.scenario}()},
\code{\link{rsim.step}()},
\code{\link{stanzaplot}()},
\code{\link{webplot}()},
\code{\link{write.Rpath}()},
\code{\link{write.Rsim}()},
\code{\link{write.rpath.params}()}
}
\concept{Rpath functions}
